2025/HW/water tank/admin/src/requirements.txt
pymodbus==3.9.2
python-dotenv